CategoryService class methods are implemented

+++ code/backend/Core/Application/services/CategoryService.cs
@@ -1,4 +1,5 @@
 using Application.interfaces;
+using Application.Utilities.Exceptions;
 using Domain;
 using Microsoft.Extensions.Logging;
 
@@ -13,16 +14,24 @@ public class CategoryService : ICategoryService {
     }
 
     public async Task<Category> AddCategoryAsync(Category category) {
-        _logger.LogInformation("CategoryService.AddCategoryAsync() method is started.");
+        _logger.LogDebug("CategoryService.AddCategoryAsync() method is started.");
+        var result = GetIfCategoryExists(category);
+        if (!result) {
+            throw new CategoryAlreadyExistsException("This category already exists");
+        }
         await _repository.AddAsync(category);
         return category;
     }
 
     public async Task<Category> DeleteCategoryAsync(int categoryId) {
-        _logger.LogInformation("CategoryService.DeleteCategoryAsync() method is started.");
+        _logger.LogDebug("CategoryService.DeleteCategoryAsync() method is started.");
         var result = GetCategoryById(categoryId);
+        if (result is null) {
+            throw new CategoryNotFoundException("Category not found");
+        }
         await _repository.DeleteAsync(result);
-        return result;//bu değişcek :)
+        return result;
+
     }
 
     public List<Category> GetCategories() {
@@ -39,11 +48,17 @@ public class CategoryService : ICategoryService {
         _logger.LogDebug("CategoryService.GetCategoryById() method is started.");
         return _repository.FindById(categoryId);
     }
+
     // ? için kontrol yapılacak
 
     public async Task<Category> UpdateCategoryAsync(Category updateCategory) {
         _logger.LogDebug("CategoryService.UpdateCategoryAsync() method is started.");
         await _repository.UpdateAsync(updateCategory);
         return updateCategory;
+    }
+
+    private bool GetIfCategoryExists(Category category) {
+        var result = _repository.Get(x => x.Id == category.Id && string.Equals(x.CategoryName, category.CategoryName, StringComparison.OrdinalIgnoreCase));
+        return result is null;
     }
 }
